🔧 Update conversion patterns in QCO, QC, and QIR dialects to ensure consistent type casting for StaticOp indices. This change enhances type safety by explicitly casting indices to int64_t during operation creation and replacement.

+++ mlir/lib/Conversion/QCOToQC/QCOToQC.cpp
@@ -26,7 +26,6 @@
 #include <mlir/Support/LogicalResult.h>
 #include <mlir/Transforms/DialectConversion.h>
 
-#include <cstdint>
 #include <utility>
 
 namespace mlir {
@@ -156,7 +155,8 @@ struct ConvertQCOStaticOp final : OpConversionPattern<qco::StaticOp> {
   matchAndRewrite(qco::StaticOp op, OpAdaptor /*adaptor*/,
                   ConversionPatternRewriter& rewriter) const override {
     // Create qc.static with the same index
-    rewriter.replaceOpWithNewOp<qc::StaticOp>(op, op.getIndex());
+    rewriter.replaceOpWithNewOp<qc::StaticOp>(
+        op, static_cast<int64_t>(op.getIndex()));
     return success();
   }
 };
